Platinum_Tier/requirements.txt
# Platinum Tier AI Employee - Python Dependencies
# ================================================
# Install with: pip install -r requirements.txt

# Core Dependencies
# -----------------
python-dotenv>=1.0.0
requests>=2.31.0

# Google API & Authentication (Gmail, Google Search)
# ---------------------------------------------------
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Odoo ERP Integration (XML-RPC)
# ------------------------------
# xmlrpc.client is part of Python standard library (no install needed)

# Social Media Management
# -----------------------
# Uses local JSON storage (no external dependencies)

# Development & Testing (Optional)
# --------------------------------
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.7.0
# flake8>=6.1.0

# Production Deployment (Optional)
# --------------------------------
# gunicorn>=21.2.0
# supervisor>=4.2.5
